
%Data manual format

\name{aves_2tree_58taxa}
\alias{aves_2tree_58taxa}

\title{Example Aves phylogenies for the mulTree package}
\description{
IN DEVELOPEMENT
}
\usage{
data(aves_2tree_58taxa)
}
\format{
The datafile contains two aves phylogenies (\code{aves_2tree_58taxa}) of 58 bird species as a 'multiPhylo' object containing two "phylo" objects . The tip names are the binomial names of the species.
\describe{
}
}
\references{Healy, K., Guillerme, T., Finlay, S., Kane, A., Kelly, S, B, A., McClean, D., Kelly, D, J., Donohue, I., Jackson, A, L., Cooper, N. (2014) Ecology and mode-of-life explain lifespan variation in birds and mammals. Proceedings of the Royal Society B 281, 20140298}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

