% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{force.graph}
\alias{force.graph}
\title{A D3 graph of a multivariate decision tree}
\usage{
force.graph(dt, ...)
}
\arguments{
\item{dt}{is \code{multree} fitted object.}
}
\value{
an interactive D3 tree object with labels.
}
\description{
This function provides an interactive interface to analyze a \code{multree} object.
}
\details{
This function builds a D3 javascript object to view and analyze a \code{multree} object. Presently there are no options to alter labels.
}
\examples{
Y  <- iris[,5]
X  <- iris[,1:4]
dt <- multree(X, Y, model = SVM)
force.graph(dt)
}
\author{
David Rodriguez
}
\seealso{
\code{networkD3}
}

