
%Data manual format

\name{lifespan.mcmc}
\alias{lifespan.mcmc}

\title{Example of mulTree analysis data}
\description{
This is an example of MCMCglmm output files using the \code{mulTree} function on the \code{lifespan} data.
}
\usage{
data(lifespan.mcmc)
}

\format{
Contains the results of a \code{mulTree} analysis on two trees with two independent chains per trees.
}

