
%Data manual format

\name{lifespan_volant_192taxa}
\alias{lifespan_volant_192taxa}

\title{Example Aves and Mammalia lifespan for the mulTree package}
\description{
IN DEVELOPEMENT
}
\usage{
data(lifespan_volant_192taxa)
}
\format{
The datafile contains a data frame (\code{lifespan_volant_192taxa}) of 192 complete cases for those species. The data frame contains four variables:
\describe{
\item{species}{The species binomial name.}
\item{longevity}{The mean centered logged maximum lifespan in years.}
\item{mass}{The mean centered logged body mass in grams.}
\item{volant}{Flying ability, as a two level factor: volant and nonvolant.}
}
}
\references{Healy, K., Guillerme, T., Finlay, S., Kane, A., Kelly, S, B, A., McClean, D., Kelly, D, J., Donohue, I., Jackson, A, L., Cooper, N. (2014) Ecology and mode-of-life explain lifespan variation in birds and mammals. Proceedings of the Royal Society B 281, 20140298}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

