% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{m.predict}
\alias{m.predict}
\title{Predicting classes from a multivariate decision tree}
\usage{
m.predict(t, X, ...)
}
\arguments{
\item{t}{is a tree from fitted by 'multree'}

\item{X}{is a dataframe with the same columns fitting 't'.}
}
\value{
a vector of class labels for each row from 'x'
}
\description{
This function predict classes from a dataframe 'x' based on a 'multree' object.
}
\details{
This function provides class predictions from a dataframe 'x' based on a 'multree' object. Nested in this prediction function are the associated predict functions for a 'glm', 'svm', 'nnet', 'lm', 'lda' object.
}
\examples{
Y  <- iris[,5]
X  <- iris[,1:4]
dt <- multree(Y, X)
p  <- m.predict(dt, X)
table(pred = p, actu = Y)
}
\author{
David Rodriguez
}
\seealso{
\code{glm, svm, nnet, lm, lda}
}

