% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forest.R
\name{mf.predict}
\alias{mf.predict}
\title{Predicting classes from a multivariate decision forest}
\usage{
mf.predict(f, X)
}
\arguments{
\item{f}{is a tree from fitted by 'mulforest'}

\item{X}{is a dataframe with the same columns fitting 'f'.}
}
\value{
a vector of class labels for each row from 'x'
}
\description{
This function predict classes from a dataframe 'x' based on a 'mulforest' object.
}
\details{
This function provides class predictions from a dataframe 'x' based on a 'mulforest' object. Nested in this prediction function are the associated predict functions for a 'glm', 'svm', 'nnet', 'lm', 'lda' object.
}
\examples{
Y  <- iris[,5]
X  <- iris[,1:4]
f  <- mulforest(Y, X)
p  <- mf.predict(f, X)
table(pred = p, actu = Y)
}
\author{
David Rodriguez
}
\seealso{
\code{glm, svm, nnet, lm, lda}
}

