% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulTree-package.R
\docType{package}
\name{mulTree-package}
\alias{mulTree-package}
\title{Performs MCMCglmm On Multiple Phylogenetic Trees.}
\description{
Allows to run a MCMCglmm on multiuple phylogenetic trees to take into account phylogenetic uncertainty.
}
\author{
Thomas Guillerme <guillert@tcd.ie> & Kevin Healy <heaylke@tcd.ie>
}
\references{
Healy, K., Guillerme T., Finlay, S., Kane, A., Kelly, S.B.A., McClean, D., Kelly, D.J., Donohue, I., Jackson, A.L. and Cooper, N. ,
2014. Ecology and mode of life explain lifespan variation in birds and mammals. Proceedings of the Royal Society of London B. 281(1784), 20140298,
}
\keyword{MCMCglmm,}
\keyword{bayesian,}
\keyword{correction,}
\keyword{distribution}
\keyword{phylogenetic}
\keyword{tree}

