% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.bind.R
\name{tree.bind}
\alias{tree.bind}
\title{Randomly binding trees together}
\usage{
tree.bind(x, y, sample, root.age)
}
\arguments{
\item{x, y}{Two \code{phylo} or \code{multiPhylo} objects.}

\item{sample}{The number of trees to create. If missing, the \code{sample} size is set to 1.}

\item{root.age}{The age of the root where both trees are combined (can be any unit). If missing, the \code{root.edge} is set to \code{0}.}
}
\value{
If \code{x}, \code{y} and \code{sample} are \eqn{>1}, the function returns a \code{multiPhylo} object; else it returns a \code{phylo} object.
}
\description{
Randomly binds trees together with a provided number of trees and a root age.
}
\examples{
## Combines 2 randomly chosen trees from x and from y into z putting the root age at 12.
x <- rmtree(10, 5) ; y <- rmtree(5, 5)
tree.bind(x, y, sample = 3, root.age = 12)

##Combines one mammal and and one bird tree and setting the root age at 250 Mya.
data(lifespan)
combined_trees <- tree.bind(trees_mammalia, trees_aves, sample = 1, root.age = 250)
plot(combined_trees) # A tree with both mammals and aves

}
\author{
Thomas Guillerme
}
