% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mulTree.R
\name{as.mulTree}
\alias{as.mulTree}
\title{Combines a data table and a "multiPhylo" object into a list to be used by the mulTree function}
\usage{
as.mulTree(data, tree, taxa, rand.terms, clean.data = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing at least two variable and taxa names.}

\item{tree}{A \code{phylo} or \code{multiPhylo} object.}

\item{taxa}{The name or the number of the column containing the list of taxa in the \code{data}.}

\item{rand.terms}{A \code{\link[stats]{formula}} ontaining additional random terms to add to the default formula (phylogenetic effect). If missing, the random terms are the column containing the taxa names and a column containing the specimen names if more than one taxa per specimen is present.}

\item{clean.data}{A \code{logical} value: whether to use the \code{\link{clean.data}} function. Default = \code{FALSE}.}
}
\value{
A \code{mulTree} object the data to be passed to the \code{\link{mulTree}} function.
}
\description{
Combines a table and a multiple phylogenies using \code{\link[caper]{comparative.data}} function. Changes the name of the taxa column into "sp.col" to be read by \code{\link[caper]{comparative.data}} and \code{\link[MCMCglmm]{MCMCglmm}}.
}
\details{
If \code{rand.terms} is specified by the user, the first element is forced to be called "animal".
}
\examples{
##Creates a data.frame
data_table <- data.frame(taxa = LETTERS[1:5], var1 = rnorm(5), var2 = c(rep("a",2), rep("b",3)))
##Creates a list of tree
tree_list <- rmtree(5,5, tip.label = LETTERS[1:5])
##Creates the "mulTree" object
as.mulTree(data_table, tree_list, taxa = "taxa")

##Creating a mulTree object with multiple specimens
##Creates a data.frame with taxa being labelled as "spec1"
data_table_sp1 <- data.frame(taxa = LETTERS[1:5], var1 = rnorm(5), var2 = c(rep("a",2), rep("b",3)), specimen = c(rep("spec1", 5)))
##Creates a data.frame with taxa being labelled as "spec2"
data_table_sp2 <- data.frame(taxa = LETTERS[1:5], var1 = rnorm(5), var2 = c(rep("a",2), rep("b",3)), specimen = c(rep("spec2", 5)))
##Combines both data.frames
data_table <- rbind(data_table_sp1, data_table_sp2)
##Creates a list of tree
tree_list <- rmtree(5,5, tip.label = LETTERS[1:5])
##Creates the "mulTree" object (with a random term formula)
as.mulTree(data_table, tree_list, taxa = "taxa", rand.terms = ~taxa+specimen)

}
\author{
Thomas Guillerme
}

