% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulTree-package.R
\docType{data}
\name{lifespan.mcmc}
\alias{lifespan.mcmc}
\title{Example of mulTree analysis data}
\format{Contains the results of a \code{mulTree} analysis on two trees with two independent chains per trees.}
\description{
This is an example of MCMCglmm output files using the \code{\link{mulTree}} function on the \code{lifespan} data.
}
\references{
Healy, K., Guillerme, T., Finlay, S., Kane, A., Kelly, S, B, A., McClean, D., Kelly, D, J., Donohue, I., Jackson, A, L., Cooper, N. (2014) Ecology and mode-of-life explain lifespan variation in birds and mammals. Proceedings of the Royal Society B 281, 20140298c
}
\keyword{datasets}
