% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulTree-package.R
\docType{data}
\name{lifespan_volant_192taxa}
\alias{lifespan_volant_192taxa}
\title{Example Aves and Mammalia lifespan for the mulTree package}
\format{The datafile contains a data frame (\code{lifespan_volant_192taxa}) of 192 complete cases for those species. The data frame contains five variables:
\describe{
       \item{species}{The species binomial name.}
       \item{class}{The species phylogenetic class.}
       \item{longevity}{The mean centred logged maximum lifespan in years.}
       \item{mass}{The mean centred logged body mass in grams.}
       \item{volant}{Flying ability, as a two level factor: volant and nonvolant.}
}}
\description{
This is a dataset containing lifespan data from 192 species of birds and mammals.
}
\references{
Healy, K., Guillerme, T., Finlay, S., Kane, A., Kelly, S, B, A., McClean, D., Kelly, D, J., Donohue, I., Jackson, A, L., Cooper, N. (2014) Ecology and mode-of-life explain lifespan variation in birds and mammals. Proceedings of the Royal Society B 281, 20140298c
}
\keyword{datasets}

