% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree.R
\name{multree}
\alias{multree}
\title{Building a multivariate decision tree}
\usage{
multree(X, Y, Pure = Gini, is.forest = FALSE, splitter = MSplit,
  model = LM, ...)
}
\arguments{
\item{X}{a dataframe of continuous or categorical values.}

\item{Y}{a vector of classes set as factors.}

\item{Pure}{may be: gini, information, twoing. Sets impurity measure for a node.}

\item{model}{may be: GLM, SVM, NET, LM, LDA. Sets the hyper surface spliter at nodes in the tree.}
}
\value{
an MulTree object with a tree consisting of nodes and their attributes.
}
\description{
This function develops a multivariate decision tree classifier in 'x' to the class of 'y'.
}
\details{
This function develops a multivariate decision tree classifier to predict an unseen 'x's class 'y'. The options of using a generalized linear model, support vector machine, neural network, or linear model to construct a hyper-surface splitting plane in ways not found in decision trees.
}
\examples{
Y  <- iris[,5]
X  <- iris[,1:4]
dt <- multree(X, Y, model = SVM)
}
\author{
David Rodriguez
}
\seealso{
\code{glm, svm, nnet, lm, lda}
}

