% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mulTree.R
\name{mulTree}
\alias{mulTree}
\title{Run MCMCglmm on multiple trees}
\usage{
mulTree(mulTree.data, formula, parameters, chains = 2, priors, ...,
  convergence = 1.1, ESS = 1000, verbose = TRUE,
  output = "mulTree_models", warn = FALSE, parallel)
}
\arguments{
\item{mulTree.data}{A list of class \code{mulTree} generated using \code{\link{as.mulTree}}.}

\item{formula}{An object of class \code{formula} (excluding the random terms).}

\item{parameters}{A list of three numerical values to be used respectively as: (1) the number of generations, (2) the sampling value, (3) the burnin.}

\item{chains}{The number of independent chains to run per model.}

\item{priors}{A series of priors to use for the MCMC. If missing, the priors will be the default parameters from the \code{\link[MCMCglmm]{MCMCglmm}} function.}

\item{...}{Any additional arguments to be passed to the \code{\link[MCMCglmm]{MCMCglmm}} function.}

\item{convergence}{A numerical value for assessing chains convergence (default = \code{1.1}).}

\item{ESS}{A numerical value for assessing the effective sample size (default = \code{1000}).}

\item{verbose}{A logical value stating whether to be verbose or not (default = \code{TRUE}).}

\item{output}{A string of characters that will be used as chain name for the models output (default = \code{mulTree_models}).}

\item{warn}{Whether to print the warning messages from the \code{\link[MCMCglmm]{MCMCglmm}} function (default = \code{FALSE}).}

\item{parallel}{An optional vector containing the virtual connection process type for running the chains in parallel (requires \code{snow} package).}
}
\value{
Generates MCMCglmm models and saves them sequentially out of \code{R} environement to minimise users RAM usage. 
Use \code{\link{read.mulTree}} to reload the models back in the \code{R} environement. 
Because of the calculation of the vcv matrix for each model and each tree in the MCMCglmm models, this function is really RAM demanding. 
For big datasets we heavily recomend to have at least 4GB RAM DDR3 available.
}
\description{
Running a \code{\link[MCMCglmm]{MCMCglmm}} model on a multiple phylogenies and a \code{data.frame} combined using \code{\link{as.mulTree}}. The results are written out of \code{R} environment as individual models.
}
\examples{
## Quick example:
## Before the analysis
data <- data.frame("sp.col" = LETTERS[1:5], var1 = rnorm(5), var2 = rnorm(5))
tree <- replicate(3, rcoal(5, tip.label = LETTERS[1:5]), simplify = FALSE)
class(tree) <- "multiPhylo"
mulTree.data <- as.mulTree(data, tree, taxa = "sp.col")
priors <- list(R = list(V = 1/2, nu = 0.002), G = list(G1 = list(V = 1/2, nu = 0.002)))
## quick example
mulTree(mulTree.data, formula = var1 ~ var2, parameters = c(10000, 10, 1000), chains = 2,
     prior = priors, output = "quick_example", convergence = 1.1, ESS = 100)
## Clean folder
file.remove(list.files(pattern = "quick_example"))
## alternative example with parallel argument (and double the number of chains!)
mulTree(mulTree.data, formula = var1 ~ var2, parameters = c(10000, 10, 1000), chains = 4,
     prior = priors, output = "quick_example", convergence = 1.1, ESS = 100, parallel = "SOCK")
## Clean folder
file.remove(list.files(pattern = "quick_example"))

\dontrun{
## Before the analysis:
## read in the data
data(lifespan)
## combine aves and mammalia trees
combined_trees <- tree.bind(x = trees_mammalia, y = trees_aves, sample = 2, root.age = 250)

## Preparing the variables for the mulTree function
## creates the "mulTree" object
mulTree_data <- as.mulTree(data = lifespan_volant, tree = combined_trees, taxa = "species")
## formula
test_formula <- longevity ~ mass + volant
## parameters (number of generations, thin/sampling, burnin)
mcmc_parameters <- c(101000, 10, 1000)
# For higher ESS run longer by increasing the number of generations
## priors
mcmc_priors <- list(R = list(V = 1/2, nu = 0.002), G = list(G1 = list(V = 1/2, nu = 0.002)))

## Running MCMCglmm on multiple trees
## WARNING: This example takes between 1 and 2 minutes to run
## and generates files in your current directory.
mulTree(mulTree_data, formula = test_formula, parameters = mcmc_parameters, priors = mcmc_priors,
     output = "longevity.example", ESS = 50)

## The models are saved out of R environment under the "longevity.example" chains names.
## Use read.mulTree() to read the generated models.

## Remove the generated files from the current directory
file.remove(list.files(pattern="longevity.example"))

## Parallel example
## Loading the snow package
library(snow)
## Running the same MCMCglmm on multiple trees
mulTree(mulTree_data, formula = test_formula, parameters = mcmc_parameters, priors = mcmc_priors,
     output = "longevity.example", ESS = 50, parallel = "SOCK")
## Remove the generated files from the current directory
file.remove(list.files(pattern = "longevity.example"))
 
## Same example but including specimens
## Subset of the data
data<-lifespan_volant[sample(nrow(lifespan_volant), 30),]
##Create a dataset with two specimen per species
data<-rbind(cbind(data, specimen = rep("spec1",30)), cbind(data, specimen = rep("spec2",30)))
##Cleaning the trees
trees<-clean.data(taxon = "species", data, combined_trees)$tree

##Creates the mulTree object
mulTree_data<-as.mulTree(data, trees, species = "species", rand.terms = ~species+specimen)

##Running MCMCglmm on multiple trees
mulTree(mulTree_data, formula = test_formula, parameters = mcmc_parameters, priors = mcmc_priors,
      output = "longevity.example", ESS = 50)
##Remove the generated files from the current directory
file.remove(list.files(pattern="longevity.example"))
}

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link[MCMCglmm]{MCMCglmm}}, \code{\link{as.mulTree}}, \code{\link{read.mulTree}}
}

