% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mulTree.R
\name{plot.mulTree}
\alias{plot.mulTree}
\title{Plots \code{mulTree} results}
\usage{
\method{plot}{mulTree}(mulTree.summary, terms, cex.terms, cex.coeff,
  horizontal = FALSE, ylim, col, ...)
}
\arguments{
\item{mulTree.summary}{A \code{mulTree} matrix summarized by \code{\link{summary.mulTree}}.}

\item{terms}{An optional vector of terms labels.}

\item{cex.terms}{An optional value for the size of the terms labels.}

\item{cex.coeff}{An optional value for the size of the coefficients labels.}

\item{horizontal}{Whether to plot the results horizontally (\code{default = FALSE}).}

\item{ylim}{Optional, the y limits of the plot.}

\item{col}{Optional, the color of the plot.}

\item{...}{Any additional arguments to be passed to \code{\link[graphics]{plot}}.}
}
\description{
Plots a boxplots of the terms of a \code{mulTree} analysis.
}
\examples{
## read in the data
data(lifespan.mcmc)

## summarising the results
summarized_data <- summary(lifespan.mcmc)

## plotting the results
plot(summarized_data)

## Same plot using more options
plot(summarized_data, horizontal = TRUE, ylab = "", ylim = c(-2,2),
     main = "Posterior distributions", cex.terms = 0.5, cex.coeff = 0.8,
     terms = c("Intercept", "BodyMass", "Volancy", "Phylogeny", "Residuals"),
     col = c("red"), cex.main = 0.8)
abline(v = 0, lty = 3)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{mulTree}}, \code{\link{read.mulTree}}, \code{\link{summary.mulTree}}
}

