% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mulTree.R
\name{read.mulTree}
\alias{read.mulTree}
\title{Reads MCMCglmm models fromn mulTree.}
\usage{
read.mulTree(mulTree.chain, convergence = FALSE, model = FALSE,
  extract = NULL)
}
\arguments{
\item{mulTree.chain}{A chain name of \code{MCMCglmm} models written by the \code{\link{mulTree}} function.}

\item{convergence}{Logical, whether to read the convergence file associated with the chain name (default = \code{FALSE}).}

\item{model}{Logical, whether to input a single \code{MCMCglmm} model or the list of random and fixed terms only (default = \code{FALSE}).}

\item{extract}{Optional, the name of one or more elements to extract from each model (rather than loading the full model; default = \code{NULL}).}
}
\value{
A \code{list} of the terms of class \code{mulTree} by default.
Else a \code{MCMCglmm} object (if \code{model = TRUE}); a \code{gelman.diag} object (if \code{convergence = TRUE}) or a list of extracted elements from the \code{MCMCglmm} models (if \code{extract} is not \code{NULL}).
}
\description{
Reads MCMCglmm objects from the \code{\link{mulTree}} function back into the \code{R} environment.
}
\details{
The argument \code{model = TRUE} can be used to load the \code{MCMCglmm} object of a unique chain.
The resulting object can be then summarized or plotted as S3 method for class \code{MCMCglmm}.
}
\examples{
## Creating some dummy mulTree models
data <- data.frame("sp.col" = LETTERS[1:5], var1 = rnorm(5), var2 = rnorm(5))
tree <- replicate(3, rcoal(5, tip.label = LETTERS[1:5]), simplify = FALSE) ; class(tree) <- "multiPhylo"
mulTree.data <- as.mulTree(data, tree, taxa = "sp.col")
priors <- list(R = list(V = 1/2, nu = 0.002), G = list(G1 = list(V = 1/2, nu = 0.002)))
mulTree(mulTree.data, formula = var1 ~ var2, parameters = c(10000, 10, 1000), chains = 2, prior = priors, output = "quick_example", convergence = 1.1, ESS = 100, verbose = FALSE)

## Reading all the models
all_chains <- read.mulTree("quick_example")
summary(all_chains)

## Reading the convergence diagnosis for all the trees
read.mulTree("quick_example", convergence = TRUE)

## Reading a specific model
model <- read.mulTree("quick_example-tree1_chain1", model = TRUE)

## Reading only the error term and the tune for all models
read.mulTree("quick_example", extract=c("error.term", "Tune"))

##Remove the generated files from the current directory
file.remove(list.files(pattern = "quick_example"))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{mulTree}}, \code{\link{plot.mulTree}}, \code{\link{summary.mulTree}}
}

