% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mulTree.R
\name{summary.mulTree}
\alias{summary.mulTree}
\title{Summarises \code{mulTree} data}
\usage{
\method{summary}{mulTree}(mulTree.results, prob = c(50, 95), use.hdr = TRUE,
  cent.tend = median, ...)
}
\arguments{
\item{mulTree.results}{A \code{mulTree} object obtained from \code{\link{read.mulTree}} function.}

\item{prob}{One or more precentage values for to be the credibility intervals (\code{default = c(50, 95)}).}

\item{use.hdr}{Logical, whether to calculate the highest density region using \code{\link[hdrcde]{hdr}} (\code{TRUE}) or the quantiles using \code{\link[stats]{quantile}} (\code{FALSE}).}

\item{cent.tend}{A function for calculating the central tendency (\code{default = median}) from the quantiles (if \code{use.hdr = FALSE}; else is ignored).}

\item{...}{Any optional arguments to be passed to the \code{\link[hdrcde]{hdr}} or \code{\link[stats]{quantile}} functions.}
}
\value{
A \code{matrix} of class \code{mulTree}.
}
\description{
Summarises the \code{MCMCglmm} models calculated from multiple trees by caculating the highest density regions (\code{\link[hdrcde]{hdr}}) of the fixed and random terms.
}
\examples{
## Read in the data
data(lifespan.mcmc)

## Summarizing all the chains
summary(lifespan.mcmc)

## Modyfing the CI
summary(lifespan.mcmc, prob = 95) == summary(lifespan.mcmc, prob = 10) #Is mainly FALSE

## Using use.hdr = FALSE
summary(lifespan.mcmc, use.hdr = FALSE)

## Difference in results
round(summary(lifespan.mcmc) - summary(lifespan.mcmc, use.hdr = FALSE), digit = 3)

## Difference in time
system.time(summary(lifespan.mcmc))
system.time(summary(lifespan.mcmc, use.hdr = FALSE))

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{mulTree}}, \code{\link{read.mulTree}}, \code{\link{plot.mulTree}}
}

