\name{sum.mulTree}
\alias{sum.mulTree}

\title{Summarise MCMCglmm data generated by \code{hdr.mulTree} function}

\description{Summarise MCMCglmm data generated by \code{hdr.mulTree} function.}

\usage{
summary.mulTree(hdr.mcmc)
}

\arguments{
\item{hdr.mcmc}{The output of \code{hdr.mulTree} function.}
}


\examples{
#Do example
}

\author{Thomas Guillerme}