% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mulTree.R
\name{summary.mulTree}
\alias{summary.mulTree}
\title{Summarises \code{mulTree} data}
\usage{
\method{summary}{mulTree}(mulTree.results, prob = c(50, 95), use.hdr = TRUE,
  cent.tend = stats::median, ...)
}
\arguments{
\item{mulTree.results}{A \code{mulTree} object obtained from \code{\link{read.mulTree}} function.}

\item{prob}{One or more precentage values for to be the credibility intervals (\code{default = c(50, 95)}).}

\item{use.hdr}{Logical, whether to calculate the highest density region using \code{\link[hdrcde]{hdr}} (\code{TRUE}) or the quantiles using \code{\link[stats]{quantile}} (\code{FALSE}).}

\item{cent.tend}{A function for calculating the central tendency (\code{default = median}) from the quantiles (if \code{use.hdr = FALSE}; else is ignored).}

\item{...}{Any optional arguments to be passed to the \code{\link[hdrcde]{hdr}} or \code{\link[stats]{quantile}} functions.}
}
\value{
A \code{matrix} of class \code{mulTree}.
}
\description{
Summarises the \code{MCMCglmm} models calculated from multiple trees by caculating the highest density regions (\code{\link[hdrcde]{hdr}}) of the fixed and random terms.
}
\details{
When using the highest density region caculation method (\code{use.hdr = TRUE}), the returned central tendency is always the first estimated mode (see \code{\link[hdrcde]{hdr}}).
Note that the results maybe vary when using \code{use.hdr = FALSE} or \code{TRUE}.
We recommend to use \code{use.hdr = TRUE} when possible.

When \code{use.hdr = FALSE}, the computation is faster but the quantiles are calculated and not estimated.
 
When \code{use.hdr = TRUE}, the computation is slower but the quantiles are estimated using the highest density regions.
The given estimates central tendency is calculated as the mode of the estimated highest density region.
For speeding up the calculations, the bandwidth (\code{h} argument) from \code{\link[hdrcde]{hdr}} can be estimated by using \code{\link[stats]{bw.nrd0}}.
}
\examples{
## Read in the data
data(lifespan.mcmc)

## Summarizing all the chains
summary(lifespan.mcmc)

## Modyfing the CI
summary(lifespan.mcmc, prob = 95)

## Using use.hdr = FALSE
summary(lifespan.mcmc, use.hdr = FALSE)

}
\author{
Thomas Guillerme
}
\seealso{
\code{\link{mulTree}}, \code{\link{read.mulTree}}, \code{\link{plot.mulTree}}
}

